Add rendering with SDL

+++ makefile
@@ -1,5 +1,16 @@
-CFLAGS=-std=c11 -pedantic -Wall -Werror -Wextra -fPIC -O3 $(sdl2-config --cflags)
-LDLIBS=-ldl $(sdl2-config --libs)
+OS=$(shell uname -s)
+
+ifeq ($(OS),Darwin)
+	OPENGL_FLAGS=
+	OPENGL_LIBS=-framework OpenGL
+else
+endif
+
+SDL_FLAGS=$(shell sdl2-config --cflags)
+SDL_LIBS=$(shell sdl2-config --libs)
+
+CFLAGS=-std=c11 -pedantic -Wall -Werror -Wextra -O3 $(SDL_FLAGS) $(OPENGL_FLAGS)
+LDLIBS=-ldl $(SDL_LIBS) $(OPENGL_LIBS)
 
 all: garden-game libgame.so
 
@@ -7,8 +18,8 @@ garden-game: main.c game.h
 	$(CC) $(CFLAGS) $(LDFLAGS) -o $@ $< $(LDLIBS)
 
 libgame.so: game.c game.h
-	$(CC) $(CFLAGS) -shared $(LDFLAGS) -o $@ $< $(LDLIBS)
-	kill -s USR1 `pgrep garden-game`
+	$(CC) $(CFLAGS) -shared $(LDFLAGS) -o $@ $< $(LDLIBS) -fPIC
+	if pgrep garden-game; then kill -s USR1 `pgrep garden-game`; fi
 
 test: garden-game libgame.so
 	./$<
